+++ Auxillary files/Update Backups/20260903-104339-generator-v0.1.4.4/js/config/projectVersion.js
@@ -8,7 +8,7 @@
 window.LyricSolitaireProject = {
     gameVersion: "0.1.0",
     simulatorVersion: "0.1.3.1",
-    generatorVersion: "0.1.4.4",
+    generatorVersion: "0.1.4.3",
     dataSchemaVersion: "1.3.1",
 
     personas: [